+++ src/dplyr.h
@@ -47,12 +47,12 @@ struct symbols {
   static SEXP dplyr_internal_signal;
   static SEXP dot_indices;
   static SEXP chops;
-  static SEXP mask;
   static SEXP vec_is_list;
   static SEXP new_env;
   static SEXP dot_data;
   static SEXP used;
   static SEXP across;
+  static SEXP env_mask_bindings;
 };
 
 struct vectors {
@@ -109,33 +109,45 @@ SEXP dplyr_summarise_recycle_chunks_in_place(SEXP list_of_chunks, SEXP list_of_r
 SEXP dplyr_group_indices(SEXP data, SEXP rows);
 SEXP dplyr_group_keys(SEXP group_data);
 
-SEXP dplyr_mask_remove(SEXP env_private, SEXP s_name);
-SEXP dplyr_mask_add(SEXP env_private, SEXP s_name, SEXP ptype, SEXP chunks);
+SEXP dplyr_mask_binding_remove(SEXP env_private, SEXP s_name);
+SEXP dplyr_mask_binding_add(SEXP env_private, SEXP s_name, SEXP ptype, SEXP chunks);
 
 SEXP dplyr_lazy_vec_chop(SEXP data, SEXP rows, SEXP ffi_grouped, SEXP ffi_rowwise);
-SEXP dplyr_data_masks_setup(SEXP chops, SEXP data, SEXP rows);
+SEXP dplyr_make_mask_bindings(SEXP chops, SEXP data);
 SEXP env_resolved(SEXP env, SEXP names);
-void add_mask_binding(SEXP name, SEXP env_bindings, SEXP env_chops);
+void add_mask_binding(SEXP name, SEXP env_mask_bindings, SEXP env_chops);
 
 SEXP dplyr_extract_chunks(SEXP df_list, SEXP df_ptype);
 
-#define DPLYR_MASK_INIT()                                                                    \
-SEXP rows = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::rows));                   \
-R_xlen_t ngroups = XLENGTH(rows);                                                            \
-SEXP caller = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::caller));               \
-SEXP mask = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::mask));                   \
-SEXP chops_env = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::chops));               \
-SEXP current_group = PROTECT(Rf_findVarInFrame(ENCLOS(chops_env), dplyr::symbols::dot_current_group)) ;\
-int* p_current_group = INTEGER(current_group);                  \
-*p_current_group = 0
+#define DPLYR_MASK_INIT()                                                                                \
+  SEXP rows = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::rows));                             \
+  R_xlen_t ngroups = XLENGTH(rows);                                                                      \
+  SEXP caller = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::caller));                         \
+  SEXP env_mask_bindings = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::env_mask_bindings));   \
+  SEXP pronoun = PROTECT(rlang::as_data_pronoun(env_mask_bindings));                                     \
+  SEXP chops_env = PROTECT(Rf_findVarInFrame(env_private, dplyr::symbols::chops));                       \
+  SEXP current_group = PROTECT(Rf_findVarInFrame(ENCLOS(chops_env), dplyr::symbols::dot_current_group)); \
+  int* p_current_group = INTEGER(current_group);                                                         \
+  *p_current_group = 0
 
 #define DPLYR_MASK_FINALISE()                                  \
-UNPROTECT(5);                                                  \
-*p_current_group = 0
+  UNPROTECT(6);                                                \
+  *p_current_group = 0
 
-#define DPLYR_MASK_SET_GROUP(INDEX) *p_current_group = INDEX + 1
+// At each iteration, we create a fresh data mask so that lexical side effects,
+// such as using `<-` in a `mutate()`, don't persist between groups
+#define DPLYR_MASK_ITERATION_INIT()                                         \
+  SEXP mask = PROTECT(rlang::new_data_mask(env_mask_bindings, R_NilValue)); \
+  Rf_defineVar(dplyr::symbols::dot_data, pronoun, mask)
 
-#define DPLYR_MASK_EVAL(quo) rlang::eval_tidy(quo, mask, caller)
+#define DPLYR_MASK_ITERATION_FINALISE()                        \
+  UNPROTECT(1)
+
+#define DPLYR_MASK_SET_GROUP(INDEX)                            \
+  *p_current_group = INDEX + 1
+
+#define DPLYR_MASK_EVAL(quo)                                   \
+  rlang::eval_tidy(quo, mask, caller)
 
 #define DPLYR_ERROR_INIT(n)                                    \
   SEXP error_data = PROTECT(Rf_allocVector(VECSXP, n));              \