+++ src/init.cpp
@@ -52,12 +52,12 @@ SEXP symbols::dplyr_internal_error = Rf_install("dplyr_internal_error");
 SEXP symbols::dplyr_internal_signal = Rf_install("dplyr_internal_signal");
 SEXP symbols::dot_indices = Rf_install(".indices");
 SEXP symbols::chops = Rf_install("chops");
-SEXP symbols::mask = Rf_install("mask");
 SEXP symbols::vec_is_list = Rf_install("vec_is_list");
 SEXP symbols::new_env = Rf_install("new.env");
 SEXP symbols::dot_data = Rf_install(".data");
 SEXP symbols::used = Rf_install("used");
 SEXP symbols::across = Rf_install("across");
+SEXP symbols::env_mask_bindings = Rf_install("env_mask_bindings");
 
 SEXP vectors::classes_vctrs_list_of = get_classes_vctrs_list_of();
 SEXP vectors::empty_int_vector = get_empty_int_vector();
@@ -111,11 +111,11 @@ static const R_CallMethodDef CallEntries[] = {
   {"dplyr_group_indices", (DL_FUNC)& dplyr_group_indices, 2},
   {"dplyr_group_keys", (DL_FUNC)& dplyr_group_keys, 1},
 
-  {"dplyr_mask_remove", (DL_FUNC)& dplyr_mask_remove, 2},
-  {"dplyr_mask_add", (DL_FUNC)& dplyr_mask_add, 4},
+  {"dplyr_mask_binding_remove", (DL_FUNC)& dplyr_mask_binding_remove, 2},
+  {"dplyr_mask_binding_add", (DL_FUNC)& dplyr_mask_binding_add, 4},
 
   {"dplyr_lazy_vec_chop_impl", (DL_FUNC)& dplyr_lazy_vec_chop, 4},
-  {"dplyr_data_masks_setup", (DL_FUNC)& dplyr_data_masks_setup, 3},
+  {"dplyr_make_mask_bindings", (DL_FUNC)& dplyr_make_mask_bindings, 2},
   {"env_resolved", (DL_FUNC)& env_resolved, 2},
 
   {"dplyr_extract_chunks", (DL_FUNC)& dplyr_extract_chunks, 2},